Modularize and move outside the file
